Revert removed test
