+++ build.sbt
@@ -158,6 +158,10 @@ lazy val monixConnect = (project in file("."))
 lazy val akka = monixConnector("akka", Dependencies.Akka)
 
 lazy val dynamodb = monixConnector("dynamodb", Dependencies.DynamoDb).aggregate(awsAuth).dependsOn(awsAuth % "compile->compile;test->test")
+  .settings(libraryDependencies ++= (CrossVersion.partialVersion(scalaVersion.value) match {
+    case Some((3, _)) => Seq.empty
+    case _ => Seq("org.mockito" %% "mockito-scala" % Versions.Mockito % Test)
+  }))
 
 lazy val hdfs = monixConnector("hdfs", Dependencies.Hdfs)
   .settings(libraryDependencies ++= (CrossVersion.partialVersion(scalaVersion.value) match {
