+++ gcs/src/it/scala/monix/connect/gcp/storage/GcsBlobSuite.scala
@@ -46,114 +46,6 @@ class GcsBlobSuite extends AsyncWordSpec with MonixTaskTest with Matchers with B
       gcsBlob.exists().asserting(_ shouldBe true)
     }
 
-    /**
-    "correctly returns some blob" in {
-      //given
-      val blobSourceOption: BlobSourceOption = mock[BlobSourceOption]
-      val googleBlob = mock[GoogleBlob]
-      when(underlying.reload(blobSourceOption)).thenReturn(googleBlob)
-
-      //when
-      val maybeBlob: Option[GcsBlob] = blob.reload(blobSourceOption).runSyncUnsafe()
-
-      //then
-      maybeBlob.isDefined shouldBe true
-      maybeBlob.get shouldBe a[GcsBlob]
-      verify(underlying, times(1)).reload(blobSourceOption)
-    }
-
-    "safely returns none whenever the underlying response was null" in {
-      //given
-      val blobSourceOption: BlobSourceOption = mock[BlobSourceOption]
-      when(underlying.reload(blobSourceOption)).thenReturn(null)
-
-      //when
-      val maybeBlob: Option[GcsBlob] = blob.reload(blobSourceOption).runSyncUnsafe()
-
-      //then
-      maybeBlob.isDefined shouldBe false
-      verify(underlying, times(1)).reload(blobSourceOption)
-    }
-  }
-
-        "implement an update operation that correctly returns some blob" in {
-      // given
-      val blobTargetOption: BlobTargetOption = mock[BlobTargetOption]
-      val googleBlob = mock[GoogleBlob]
-      when(underlying.update(blobTargetOption)).thenReturn(googleBlob)
-
-      //when
-      val maybeBlob: GcsBlob = blob.update(blobTargetOption).runSyncUnsafe()
-
-      //then
-      maybeBlob shouldBe a[GcsBlob]
-      verify(underlying, times(1)).update(blobTargetOption)
-    }
-
-      "implement an copy operation" that {
-
-      "copies this blob to the target blob id" in {
-        // given
-        val blobSourceOption: BlobSourceOption = mock[BlobSourceOption]
-        val blobId = genBlobId.sample.get
-        val copywriter = mock[CopyWriter]
-        when(underlying.copyTo(blobId, blobSourceOption)).thenReturn(copywriter)
-
-        //when
-        val maybeBlob: GcsBlob = blob.copyTo(blobId, blobSourceOption).runSyncUnsafe()
-
-        //then
-        maybeBlob shouldBe a[GcsBlob]
-        verify(underlying, times(1)).copyTo(blobId, blobSourceOption)
-      }
-
-      "copies this blob to the target bucket" in {
-        // given
-        val blobSourceOption: BlobSourceOption = mock[BlobSourceOption]
-        val copywriter = mock[CopyWriter]
-        val bucketName = genNonEmtyStr.sample.get
-        val blobName = genNonEmtyStr.sample.get
-        when(underlying.copyTo(bucketName, blobSourceOption)).thenReturn(copywriter)
-
-        //when
-        val maybeBlob: GcsBlob = blob.copyTo(bucketName, blobSourceOption).runSyncUnsafe()
-
-        //then
-        maybeBlob shouldBe a[GcsBlob]
-        verify(underlying, times(1)).copyTo(bucketName, blobSourceOption)
-      }
-
-      "copies this blob to the target blob in the target bucket" in {
-        // given
-        val blobSourceOption: BlobSourceOption = mock[BlobSourceOption]
-        val copywriter = mock[CopyWriter]
-        val bucketName = genNonEmtyStr.sample.get
-        val blobName = genNonEmtyStr.sample.get
-        when(underlying.copyTo(bucketName, blobName, blobSourceOption)).thenReturn(copywriter)
-
-        //when
-        val maybeBlob: GcsBlob = blob.copyTo(bucketName, blobName, blobSourceOption).runSyncUnsafe()
-
-        //then
-        maybeBlob shouldBe a[GcsBlob]
-        verify(underlying, times(1)).copyTo(bucketName, blobName, blobSourceOption)
-      }
-    }
-
-        // "implement an async sign url operation that correctly returns some url" in {
-    //   // given
-    //   val signUrlOption: SignUrlOption = mock[SignUrlOption]
-    //   val url = new URL("TCP")
-    //   when(underlying.signUrl(2, TimeUnit.MINUTES, signUrlOption)).thenReturn(url)
-    //
-    //   //when
-    //   val maybeUrl: URL = blob.signUrl(2.minutes, signUrlOption).runSyncUnsafe()
-    //
-    //   maybeUrl shouldBe a[URL]
-    //   verify(underlying, times(1)).signUrl(2, TimeUnit.MINUTES, signUrlOption)
-    // }
-      **/
-
     "return delete if exists" in {
       val blobPath = Gen.identifier.sample.get
       val blobInfo: BlobInfo = BlobInfo.newBuilder(BlobId.of(testBucketName, blobPath)).build
@@ -335,89 +227,7 @@ class GcsBlobSuite extends AsyncWordSpec with MonixTaskTest with Matchers with B
       r2 shouldBe groupAcl
       l should contain theSameElementsAs List(userAcl, groupAcl)
     }
-
-       "implement a create acl operation that correctly returns some acl" in {
-      // given
-      val acl = genAcl.sample.get
-      when(underlying.createAcl(acl)).thenReturn(acl)
-
-      //when
-      val maybeAcl: Acl = blob.createAcl(acl).runSyncUnsafe()
-
-      //then
-      maybeAcl shouldBe a[Acl]
-      verify(underlying, times(1)).createAcl(acl)
-    }
-
-    "implement a get acl operation" that {
-
-      "that safely returns none whenever the underlying response was null" in {
-        // given
-        val acl = mock[Acl.Entity]
-        when(underlying.getAcl(acl)).thenReturn(null)
-
-        //when
-        val maybeAcl: Option[Acl] = blob.getAcl(acl).runSyncUnsafe()
-
-        //then
-        maybeAcl.isDefined shouldBe false
-        verify(underlying, times(1)).getAcl(acl)
-      }
-
-      "that correctly returns some acl" in {
-        // given
-        val acl = genAcl.sample.get
-        when(underlying.getAcl(acl.getEntity)).thenReturn(acl)
-
-        //when
-        val maybeAcl: Option[Acl] = blob.getAcl(acl.getEntity).runSyncUnsafe()
-
-        //then
-        maybeAcl.isDefined shouldBe true
-        maybeAcl.get shouldBe a[Acl]
-        verify(underlying, times(1)).getAcl(acl.getEntity)
-      }
-    }
-
-    "implement a update operation that correctly returns some acl" in {
-      // given
-      val acl = genAcl.sample.get
-      when(underlying.updateAcl(acl)).thenReturn(acl)
-
-      //when
-      val maybeAcl: Acl = blob.updateAcl(acl).runSyncUnsafe()
-
-      //then
-      maybeAcl shouldBe a[Acl]
-      verify(underlying, times(1)).updateAcl(acl)
-    }
-
-    "implement a delete operation that deletes the specified acl" in {
-      // given
-      val acl = mock[Acl.Entity]
-      when(underlying.deleteAcl(acl)).thenReturn(true)
-
-      //when
-      val result: Boolean = blob.deleteAcl(acl).runSyncUnsafe()
-
-      //then
-      result shouldBe true
-      verify(underlying, times(1)).deleteAcl(acl)
-    }
-
-    "implement a list acl operation that correctly returns zero or more acls" in {
-      // given
-      val acls = Gen.nonEmptyListOf(genAcl).sample.get
-      when(underlying.listAcls()).thenReturn(acls.asJava)
-
-      //when
-      val result: List[Acl] = blob.listAcls().toListL.runSyncUnsafe()
-
-      //then
-      result shouldBe acls
-      verify(underlying, times(1)).listAcls()
-    }
-      */
+     */
 
   }
